refactor(agentPresenter): move loop module
