+++ src/main/presenter/agentPresenter/loop/agentLoopHandler.ts
