+++ src/main/presenter/agentPresenter/loop/errorClassification.ts
