+++ src/main/presenter/agentPresenter/loop/index.ts
@@ -1,1 +1,6 @@
-export {}
+export { AgentLoopHandler } from './agentLoopHandler'
+export { LoopOrchestrator } from './loopOrchestrator'
+export { ToolCallHandler } from './toolCallHandler'
+export { ToolCallProcessor } from './toolCallProcessor'
+export { isNonRetryableError } from './errorClassification'
+export type { LoopState, StreamState } from './loopState'
