+++ src/main/presenter/agentPresenter/loop/loopState.ts
