+++ src/main/presenter/agentPresenter/loop/toolCallProcessor.ts
