+++ src/main/presenter/sessionPresenter/index.ts
@@ -24,7 +24,7 @@ import type { ISessionPresenter } from './interface'
 import { MessageManager } from './managers/messageManager'
 import type { GeneratingMessageState } from './streaming/types'
 import { ContentBufferHandler } from './streaming/contentBufferHandler'
-import { ToolCallHandler } from './loop/toolCallHandler'
+import { ToolCallHandler } from '../agentPresenter/loop/toolCallHandler'
 import { LLMEventHandler } from './streaming/llmEventHandler'
 import { SearchHandler } from '../searchPresenter/handlers/searchHandler'
 import { StreamGenerationHandler } from './streaming/streamGenerationHandler'
